backend/requirements.txt
amqp==5.0.2
appdirs==1.4.3
asgiref==3.3.1
billiard==3.6.3.0
CacheControl==0.12.6
celery==5.0.5
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-celery-beat==2.1.0
django-cors-headers==3.6.0
django-rest-framework==0.1.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
drf-yasg==1.20.0
html5lib==1.0.1
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
prompt-toolkit==3.0.8
pycparser==2.20
pyparsing==2.4.6
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
pytoml==0.1.21
pytz==2020.5
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.14.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
